requirements.txt
anthropic
discord.py
python-dotenv==1.0.0
